peer-manager: split snapshot helpers
